requirements.txt
latta-python-recorder
python-dotenv